refactor(#35): Split dropdown into standalone component
